pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "probablyprofit"
version = "1.1.0"
description = "AI-powered prediction market trading bots. Write strategy in English, let AI trade for you."
authors = [{name = "ProbablyProfit Contributors", email = "hello@probablyprofit.dev"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = [
    "polymarket",
    "prediction-markets",
    "trading-bot",
    "ai",
    "llm",
    "gpt",
    "claude",
    "automated-trading",
    "crypto",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

# Core dependencies - minimal install
dependencies = [
    # CLI & UX
    "click>=8.1.0",
    "rich>=13.0.0",
    # HTTP & async
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    # Logging
    "loguru>=0.7.0",
    # Resilience
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
# AI Providers - install what you need
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.39.0"]
google = ["google-genai>=1.0.0"]

# All AI providers
ai = [
    "openai>=1.0.0",
    "anthropic>=0.39.0",
    "google-genai>=1.0.0",
]

# Polymarket trading
polymarket = [
    "eth-account>=0.11.0",
    "web3>=6.0.0",
    "py-clob-client>=0.20.0",
    "websockets>=12.0",
]

# Intelligence layer (news, sentiment)
intel = [
    "tweepy>=4.14.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
]

# Data & backtesting
data = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

# Database persistence
db = [
    "sqlmodel>=0.0.14",
    "alembic>=1.12.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",  # Required for async SQLAlchemy
]

# Full install - everything
full = [
    "probablyprofit[ai,polymarket,intel,data,db]",
]

# Development
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.7.0",
    "bandit[toml]>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/randomness11/probablyprofit"
Documentation = "https://github.com/randomness11/probablyprofit#readme"
Repository = "https://github.com/randomness11/probablyprofit"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[project.scripts]
probablyprofit = "probablyprofit.cli.main:main"
pp = "probablyprofit.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["probablyprofit*"]
exclude = ["frontend*", "tests*"]

[tool.setuptools.package-data]
probablyprofit = ["py.typed"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "build", "dist"]
skips = ["B101", "B104"]  # Skip assert warnings (B101) and hardcoded binding (B104)

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["probablyprofit/tests"]
timeout = 30
markers = [
    "slow: marks tests as slow (run with --run-slow)",
    "integration: integration tests requiring external services",
]
filterwarnings = [
    "ignore::DeprecationWarning:websockets.*",
    "ignore::pydantic.PydanticDeprecatedSince20",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # Unused function argument (common in base classes)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]
"__init__.py" = ["F401"]
